refactor(schema): move match and json capture schema to common file
